improve search module, cleaned up console logs
