+++ src/components/common/customSearchModule/CustomSearchModule.module.scss
